Move print function to more apporpirate location
