+++ link-grammar/print/print.h
@@ -27,7 +27,6 @@ void   print_sentence_word_alternatives(Sentence sent, bool debugprint,
        void (*display)(Dictionary, const char *), struct tokenpos *);
 void print_with_subscript_dot(const char *);
 void print_chosen_disjuncts_words(const Linkage);
-void print_lwg_path(Gword **);
 
 // Used for debug/error printing
 void print_sentence_context(Sentence, dyn_str*);
