+++ link-grammar/tokenize/wordgraph.h
@@ -22,6 +22,7 @@ size_t gwordlist_len(const Gword **);
 void gwordlist_append(Gword ***, Gword *);
 void gwordlist_append_list(const Gword ***, const Gword **);
 void gword_set_print(const gword_set *);
+void print_lwg_path(Gword **);
 
 const Gword **wordgraph_hier_position(Gword *);
 void print_hier_position(const Gword *);
